refactor(api): cleanup module list command
Refs: #57
